link extractor to ledger-service/cli-opts properly
