+++ extractor/BUILD.bazel
@@ -91,6 +91,7 @@ da_scala_library(
         "//language-support/scala/bindings",
         "//ledger-api/rs-grpc-akka",
         "//ledger-api/rs-grpc-bridge",
+        "//ledger-service/cli-opts",
         "//ledger-service/lf-value-json",
         "//ledger-service/utils",
         "//ledger/ledger-api-client",
