+++ ledger-service/cli-opts/BUILD.bazel
@@ -5,16 +5,17 @@ load(
     "//bazel_tools:scala.bzl",
     "da_scala_library",
     "lf_scalacopts",
-
+)
 
 da_scala_library(
     name = "cli-opts",
-    scalacopts = lf_scalacopts,
     srcs = glob(["src/main/scala/**/*.scala"]),
-    tags = ["maven_coordinates=com.daml:http-json-utils:__VERSION__"],
+    scalacopts = lf_scalacopts,
+    tags = ["maven_coordinates=com.daml:http-json-cli-opts:__VERSION__"],
     visibility = ["//visibility:public"],
     deps = [
         "//ledger/ledger-api-common",
         "@maven//:com_github_scopt_scopt_2_12",
+        "@maven//:io_netty_netty_handler",
     ],
 )
