+++ ledger-service/cli-opts/src/main/scala/ledger/api/tls/TlsConfigurationCli.scala
@@ -1,15 +1,13 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.extractor.config
-
-import com.daml.ledger.api.tls.TlsConfiguration
+package com.daml.ledger.api.tls
 
 import java.nio.file.Paths
 
 import scala.util.Try
 
-object TlsConfigurationParser {
+object TlsConfigurationCli {
   def parse[C](parser: scopt.OptionParser[C], colSpacer: String)(
       setter: (TlsConfiguration => TlsConfiguration, C) => C): Unit = {
     def enableSet(tlsUp: TlsConfiguration => TlsConfiguration, c: C) =
