+++ release/artifacts.yaml
@@ -120,6 +120,8 @@
   type: jar-scala
 - target: //ledger-service/lf-value-json:lf-value-json
   type: jar-scala
+- target: //ledger-service/cli-opts:cli-opts
+  type: jar-scala
 - target: //ledger-service/jwt:jwt
   type: jar-scala
 - target: //ledger-service/utils:utils
